WIN32 - implemented more gamepad functionality

+++ include/ccore/gamepad.h
@@ -50,7 +50,7 @@ typedef struct {
 typedef struct {
 	char *name;
 	bool *button;
-	double *axis;
+	int *axis;
 
 	int axisAmount, buttonAmount, id;
